pyproject.toml
[project]
name = "backend-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "CaoDangHai",email = "caodanghai.it@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn[standard] (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "nltk (>=3.9.4,<4.0.0)",
    "tensorflow (>=2.21.0,<3.0.0)",
    "langchain-google-genai (>=4.2.2,<5.0.0)",
    "pydantic[email] (>=2.13.2,<3.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "langchain-groq (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
